Move config generation routes to another file
