+++ ydb/core/blobstorage/nodewarden/distconf.h
@@ -313,6 +313,8 @@ namespace NKikimr::NStorage {
         void ProcessCollectConfigs(TEvGather::TCollectConfigs *res);
         void ProcessProposeStorageConfig(TEvGather::TProposeStorageConfig *res);
 
+        struct TExConfigError : yexception {};
+
         bool GenerateFirstConfig(NKikimrBlobStorage::TStorageConfig *config);
         void AllocateStaticGroup(NKikimrBlobStorage::TStorageConfig *config);
         bool UpdateConfig(NKikimrBlobStorage::TStorageConfig *config);
